backend/requirements.txt
flask
flask-socketio
flask-mail
flask_sqlalchemy
flask_migrate
flask-jwt-extended
flask-cors
python-dotenv
gunicorn
psycopg2-binary
pytest
